// src/config/interfaces/environmets.interface.ts
export interface TKEnvironments {
  SECRETIDENTIFIER?: string;
  SECRETHIDDENKEY?: string;
  PRONOSBEURL?: string;
  PROY_CRM_BD?: string;
  PORT?: string;
  MODE?: string;
  PROY_BD?: string;
  PROY_BD_HOST?: string;
  PROY_BD_USER?: string;
  PROY_BD_PASS?: string;
  PROY_BD_PORT?: string;
  SECRET_HIDDEN_KEY?: string;
  SEND_TRANSPORTER?: string;
  SENDER?: string;
  SENDER_PASS?: string;
  SENDER_ALIAS?: string;
  DIR: string;
  CLIENT_ID?: string;
  SENDGRID_KEY?: string;
  PROY_BEURL?: string;
  PROY_FEURL?: string;
  PROY_EXTERNAL?: string;
  DIR_ASSETS?: string;

  S3_ACCESS_KEY_ID?: string;
  S3_SECRET_ACCESS_KEY?: string;
  S3_ENDPOINT_URL?: string;
  S3_BUCKET_NAME?: string;

  ONE_SIGNAL_APP_ID?: string;
  ONE_SIGNAL_API_KEY?: string;

  //SENGRID
  SENGRID_KEY?: string;
}